+++ _docs/aleph.config.ts
@@ -1,5 +1,5 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
-import markdown from "https://deno.land/x/aleph/plugins/markdown.ts";
+import markdown from "./plugins/markdown.ts";
 import type { Config } from "aleph/types";
 import unocss from "./plugins/unocss.ts";
 
@@ -9,8 +9,9 @@ export default <Config> {
     markdown({
       highlight: {
         provider: "highlight.js",
-        theme: "nord",
+        theme: "rainbow",
       },
+      toc: [2, 3],
     }),
   ],
   css: {