+++ _docs/app.tsx
@@ -1,6 +1,5 @@
 import React, { ComponentType } from "react";
 import Header from "@component/header.tsx";
-import Footer from "@component/footer.tsx";
 
 import "https://esm.sh/@unocss/reset/tailwind.css";
 
@@ -11,12 +10,18 @@ export default function App(
     <>
       <head>
         <link rel="icon" href="/logo.svg" type="image/svg+xml" />
+        <style>
+          {`
+        html, body {
+          scroll-behavior: smooth;
+          -webkit-font-smoothing: antialiased;
+          -moz-osx-font-smoothing: grayscale;
+        }
+      `}
+        </style>
       </head>
-      <main className="text-gray-700 antialiased">
-        <Header />
-        <Page {...pageProps} />
-        <Footer />
-      </main>
+      <Header />
+      <Page {...pageProps} />
     </>
   );
 }