+++ _docs/components/docs_footer.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function Footer() {
+  return (
+    <footer className="pb-30 px-4 lg:px-8 mt-6">
+      <span className="text-gray-400">
+        Copyright © 2021 TomokiMiyauci
+      </span>
+    </footer>
+  );
+}