+++ dummy/any.ts
@@ -2,18 +2,15 @@
 // This module is browser compatible.
 
 import { equality } from "../helper/equal.ts";
-import { isFunction, isObject, isSymbol } from "../deps.ts";
+import { isObject, isSymbol } from "../deps.ts";
 import type { Equality } from "../helper/equal.ts";
 import { isAnyNumber } from "./any_number.ts";
 import { isAnyString } from "./any_string.ts";
 import { isAnyBoolean } from "./any_boolean.ts";
+import { isAnyFunction } from "./any_function.ts";
 
 function isBigInt(value: unknown): value is bigint {
   return typeof value === "bigint";
-}
-
-function isAnyFunction(value: unknown): value is Function {
-  return isFunction(value) || value instanceof Function;
 }
 
 function isAnyBigInt(value: unknown): value is bigint | BigInt {