+++ dummy/mod.ts
@@ -12,3 +12,4 @@ export { stringMatching } from "./string_matching.ts";
 export { stringContaining } from "./string_containing.ts";
 export { anyOf } from "./any_of.ts";
 export { anyArray } from "./any_array.ts";
+export { anyFunction } from "./any_function.ts";